+++ src/JK.Common.Abstractions/INameableIdentifiable.cs
@@ -1,5 +1,0 @@
-﻿namespace JK.Common;
-
-public interface INameableIdentifiable<T> : IIdentifiable<T>, INameable
-{
-}